feat: Put links to README auth guides into errors (#318)
* chore: Put links to README auth guides into errors

For quicker setup process, the exact links to README auth guides, and
steps were put right into error messages, so user can open them without
reopening the repo if they closed it, or searching for it for the first
time. It is also nice to save user time on searching for needed guide in
README's text.

* style: inline return in auth_token_readme_link

* style: Remove redundant return from auth_token_readme_link